refactor: move some server files around
