+++ server/src/game/map.ts
@@ -1,22 +1,22 @@
-import { type MapDef, MapDefs } from "../../shared/defs/mapDefs";
-import { MapObjectDefs } from "../../shared/defs/mapObjectDefs";
+import { type MapDef, MapDefs } from "../../../shared/defs/mapDefs";
+import { MapObjectDefs } from "../../../shared/defs/mapObjectDefs";
 import type {
     BuildingDef,
     ObstacleDef,
     StructureDef
-} from "../../shared/defs/mapObjectsTyping";
-import { GameConfig } from "../../shared/gameConfig";
-import { MapMsg } from "../../shared/msgs/mapMsg";
-import { MsgStream, MsgType } from "../../shared/net";
-import { type AABB, type Collider, coldet } from "../../shared/utils/coldet";
-import { collider } from "../../shared/utils/collider";
-import { mapHelpers } from "../../shared/utils/mapHelpers";
-import { math } from "../../shared/utils/math";
-import { ObjectType } from "../../shared/utils/objectSerializeFns";
-import type { River } from "../../shared/utils/river";
-import { type MapRiverData, generateTerrain } from "../../shared/utils/terrainGen";
-import { util } from "../../shared/utils/util";
-import { type Vec2, v2 } from "../../shared/utils/v2";
+} from "../../../shared/defs/mapObjectsTyping";
+import { GameConfig } from "../../../shared/gameConfig";
+import { MapMsg } from "../../../shared/msgs/mapMsg";
+import { MsgStream, MsgType } from "../../../shared/net";
+import { type AABB, type Collider, coldet } from "../../../shared/utils/coldet";
+import { collider } from "../../../shared/utils/collider";
+import { mapHelpers } from "../../../shared/utils/mapHelpers";
+import { math } from "../../../shared/utils/math";
+import { ObjectType } from "../../../shared/utils/objectSerializeFns";
+import type { River } from "../../../shared/utils/river";
+import { type MapRiverData, generateTerrain } from "../../../shared/utils/terrainGen";
+import { util } from "../../../shared/utils/util";
+import { type Vec2, v2 } from "../../../shared/utils/v2";
 import type { Game } from "./game";
 import { Building } from "./objects/building";
 import { Obstacle } from "./objects/obstacle";
