+++ server/src/game/objects/building.ts
@@ -1,16 +1,16 @@
-import { MapObjectDefs } from "../../../shared/defs/mapObjectDefs";
+import { MapObjectDefs } from "../../../../shared/defs/mapObjectDefs";
 import type {
     BuildingDef,
     ObstacleDef,
     StructureDef
-} from "../../../shared/defs/mapObjectsTyping";
-import { Puzzles } from "../../../shared/defs/puzzles";
-import { type AABB, type Collider, coldet } from "../../../shared/utils/coldet";
-import { collider } from "../../../shared/utils/collider";
-import { mapHelpers } from "../../../shared/utils/mapHelpers";
-import { math } from "../../../shared/utils/math";
-import { ObjectType } from "../../../shared/utils/objectSerializeFns";
-import { type Vec2, v2 } from "../../../shared/utils/v2";
+} from "../../../../shared/defs/mapObjectsTyping";
+import { Puzzles } from "../../../../shared/defs/puzzles";
+import { type AABB, type Collider, coldet } from "../../../../shared/utils/coldet";
+import { collider } from "../../../../shared/utils/collider";
+import { mapHelpers } from "../../../../shared/utils/mapHelpers";
+import { math } from "../../../../shared/utils/math";
+import { ObjectType } from "../../../../shared/utils/objectSerializeFns";
+import { type Vec2, v2 } from "../../../../shared/utils/v2";
 import type { Game } from "../game";
 import { getColliders } from "../map";
 import type { Decal } from "./decal";
