+++ server/src/game/objects/bullet.ts
@@ -1,14 +1,14 @@
-import { GameObjectDefs } from "../../../shared/defs/gameObjectDefs";
-import type { BulletDef } from "../../../shared/defs/gameObjects/bulletDefs";
-import { MapObjectDefs } from "../../../shared/defs/mapObjectDefs";
-import type { ObstacleDef } from "../../../shared/defs/mapObjectsTyping";
-import { GameConfig } from "../../../shared/gameConfig";
-import { coldet } from "../../../shared/utils/coldet";
-import { collider } from "../../../shared/utils/collider";
-import { math } from "../../../shared/utils/math";
-import { ObjectType } from "../../../shared/utils/objectSerializeFns";
-import { util } from "../../../shared/utils/util";
-import { type Vec2, v2 } from "../../../shared/utils/v2";
+import { GameObjectDefs } from "../../../../shared/defs/gameObjectDefs";
+import type { BulletDef } from "../../../../shared/defs/gameObjects/bulletDefs";
+import { MapObjectDefs } from "../../../../shared/defs/mapObjectDefs";
+import type { ObstacleDef } from "../../../../shared/defs/mapObjectsTyping";
+import { GameConfig } from "../../../../shared/gameConfig";
+import { coldet } from "../../../../shared/utils/coldet";
+import { collider } from "../../../../shared/utils/collider";
+import { math } from "../../../../shared/utils/math";
+import { ObjectType } from "../../../../shared/utils/objectSerializeFns";
+import { util } from "../../../../shared/utils/util";
+import { type Vec2, v2 } from "../../../../shared/utils/v2";
 import type { Game } from "../game";
 import type { DamageParams, GameObject } from "./gameObject";
 import { Obstacle } from "./obstacle";
