+++ server/src/game/objects/obstacle.ts
@@ -1,14 +1,14 @@
-import { GameObjectDefs } from "../../../shared/defs/gameObjectDefs";
-import { MapObjectDefs } from "../../../shared/defs/mapObjectDefs";
-import type { ObstacleDef } from "../../../shared/defs/mapObjectsTyping";
-import * as net from "../../../shared/net";
-import { type Collider, coldet } from "../../../shared/utils/coldet";
-import { collider } from "../../../shared/utils/collider";
-import { mapHelpers } from "../../../shared/utils/mapHelpers";
-import { math } from "../../../shared/utils/math";
-import { ObjectType } from "../../../shared/utils/objectSerializeFns";
-import { util } from "../../../shared/utils/util";
-import { type Vec2, v2 } from "../../../shared/utils/v2";
+import { GameObjectDefs } from "../../../../shared/defs/gameObjectDefs";
+import { MapObjectDefs } from "../../../../shared/defs/mapObjectDefs";
+import type { ObstacleDef } from "../../../../shared/defs/mapObjectsTyping";
+import * as net from "../../../../shared/net";
+import { type Collider, coldet } from "../../../../shared/utils/coldet";
+import { collider } from "../../../../shared/utils/collider";
+import { mapHelpers } from "../../../../shared/utils/mapHelpers";
+import { math } from "../../../../shared/utils/math";
+import { ObjectType } from "../../../../shared/utils/objectSerializeFns";
+import { util } from "../../../../shared/utils/util";
+import { type Vec2, v2 } from "../../../../shared/utils/v2";
 import type { Game } from "../game";
 import type { Building } from "./building";
 import { BaseGameObject, type DamageParams } from "./gameObject";
