+++ server/src/game/objects/projectile.ts
@@ -1,12 +1,12 @@
-import { GameObjectDefs } from "../../../shared/defs/gameObjectDefs";
-import type { ThrowableDef } from "../../../shared/defs/gameObjects/throwableDefs";
-import { GameConfig } from "../../../shared/gameConfig";
-import { type Collider, coldet } from "../../../shared/utils/coldet";
-import { collider } from "../../../shared/utils/collider";
-import { math } from "../../../shared/utils/math";
-import { ObjectType } from "../../../shared/utils/objectSerializeFns";
-import { util } from "../../../shared/utils/util";
-import { type Vec2, v2 } from "../../../shared/utils/v2";
+import { GameObjectDefs } from "../../../../shared/defs/gameObjectDefs";
+import type { ThrowableDef } from "../../../../shared/defs/gameObjects/throwableDefs";
+import { GameConfig } from "../../../../shared/gameConfig";
+import { type Collider, coldet } from "../../../../shared/utils/coldet";
+import { collider } from "../../../../shared/utils/collider";
+import { math } from "../../../../shared/utils/math";
+import { ObjectType } from "../../../../shared/utils/objectSerializeFns";
+import { util } from "../../../../shared/utils/util";
+import { type Vec2, v2 } from "../../../../shared/utils/v2";
 import type { Game } from "../game";
 import { BaseGameObject } from "./gameObject";
