+++ server/src/utils/serverHelpers.ts
@@ -1,0 +1,75 @@
+import type { HttpResponse } from "uWebSockets.js";
+
+/**
+ * Apply CORS headers to a response.
+ * @param res The response sent by the server.
+ */
+export function cors(res: HttpResponse): void {
+    res.writeHeader("Access-Control-Allow-Origin", "*")
+        .writeHeader("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS")
+        .writeHeader(
+            "Access-Control-Allow-Headers",
+            "origin, content-type, accept, x-requested-with"
+        )
+        .writeHeader("Access-Control-Max-Age", "3600");
+}
+
+export function forbidden(res: HttpResponse): void {
+    res.writeStatus("403 Forbidden").end("403 Forbidden");
+}
+
+export function returnJson(res: HttpResponse, data: Record<string, unknown>): void {
+    res.writeHeader("Content-Type", "application/json").end(JSON.stringify(data));
+}
+
+/**
+ * Read the body of a POST request.
+ * @link https://github.com/uNetworking/uWebSockets.js/blob/master/examples/JsonPost.js
+ * @param res The response from the client.
+ * @param cb A callback containing the request body.
+ * @param err A callback invoked whenever the request cannot be retrieved.
+ */
+export function readPostedJSON<T>(
+    res: HttpResponse,
+    cb: (json: T) => void,
+    err: () => void
+): void {
+    let buffer: Buffer | Uint8Array;
+    /* Register data cb */
+    res.onData((ab, isLast) => {
+        const chunk = Buffer.from(ab);
+        if (isLast) {
+            let json: T;
+            if (buffer) {
+                try {
+                    // @ts-expect-error JSON.parse can accept a Buffer as an argument
+                    json = JSON.parse(Buffer.concat([buffer, chunk]));
+                } catch (_e) {
+                    /* res.close calls onAborted */
+                    res.close();
+                    return;
+                }
+                cb(json);
+            } else {
+                try {
+                    // @ts-expect-error JSON.parse can accept a Buffer as an argument
+                    json = JSON.parse(chunk);
+                } catch (_e) {
+                    /* res.close calls onAborted */
+                    res.close();
+                    return;
+                }
+                cb(json);
+            }
+        } else {
+            if (buffer) {
+                buffer = Buffer.concat([buffer, chunk]);
+            } else {
+                buffer = Buffer.concat([chunk]);
+            }
+        }
+    });
+
+    /* Register error cb */
+    res.onAborted(err);
+}
